Feature/rename mac bundle (#2062)
* Rename bundle from Modumate-Mac-Shipping to Modumate

* add custom URL scheme and microphone to plist

* Add IS_JENKINS flag

* fix rpaths in UE4 dylibs

* Fix/mac rpaths (#2061)

* app bundle being over-copied

* don't rename binary

* dump params

* path dumpery

* brute force experiment

* try not adding rpaths

* cd to relative path for rpath set

* manually strip epic's rpaths

* more elegant phys3

* quotes

* bad spaces

* lose environment variable for epic phys path

* try restore app rename

+++ Modumate/Scripts/MacPostBuild.sh
@@ -14,6 +14,8 @@ if [[ -z "$IDENTITY" ]]; then
     IDENTITY=$MODUMATE_ENV_IDENTITY
 fi
 
+IS_JENKINS=$6
+
 oldDir=$(pwd)
 highlight=$(tput setaf 2)
 normal=$(tput sgr0)
@@ -80,25 +82,67 @@ cp -R "${CEF_FRAMEWORK}" "${UE_OUTPUT_APP}/Contents/Frameworks/UnrealCEFSubProce
 mkdir -p "${UE_OUTPUT_APP}/Contents/Frameworks/UnrealCEFSubProcess (Renderer).app/Contents/Frameworks"
 cp -R "${CEF_FRAMEWORK}" "${UE_OUTPUT_APP}/Contents/Frameworks/UnrealCEFSubProcess (Renderer).app/Contents/Frameworks"
 
-#-------------------------------------
-# Do some cleanup of unused components
-if [[ -d "${UE_OUTPUT_APP}/Contents/UE4/Engine/Build" ]]; then
-    echo "${highlight}Removing Contents/UE4/Engine/Build${normal}"
-    rm -rf "${UE_OUTPUT_APP}/Contents/UE4/Engine/Build"
-fi
-if [[ -d "${UE_OUTPUT_APP}/Contents/Resources/RadioEffectUnit.component" ]]; then
-    echo "${highlight}Removing CContents/Resources/RadioEffectUnit.component${normal}"
-    rm -rf "${UE_OUTPUT_APP}/Contents/Resources/RadioEffectUnit.component"
-fi
-
-#-------------------------------------
-# Update the rpaths
 INFO_PLIST="${UE_OUTPUT_APP}/Contents/Info.plist"
 
+#-------------------------------------
+# For shipping, we need to strip the 
+# Mac-Shipping part of the bundle
+# name before any of the following
+# steps
+if [[ ! -z $IS_JENKINS ]]; then
+
+    #-------------------------------------
+    # Do some cleanup of unused components
+    if [[ -d "${UE_OUTPUT_APP}/Contents/UE4/Engine/Build" ]]; then
+        echo "${highlight}Removing Contents/UE4/Engine/Build${normal}"
+        rm -rf "${UE_OUTPUT_APP}/Contents/UE4/Engine/Build"
+    fi
+    if [[ -d "${UE_OUTPUT_APP}/Contents/Resources/RadioEffectUnit.component" ]]; then
+        echo "${highlight}Removing CContents/Resources/RadioEffectUnit.component${normal}"
+        rm -rf "${UE_OUTPUT_APP}/Contents/Resources/RadioEffectUnit.component"
+    fi
+
+    echo "Changing Modumate-Mac-Shipping.app to Modumate.app"
+    folder=$(dirname "${UE_OUTPUT_APP}")
+    
+    #move the app bundle from Modumate-Mac-Shipping.app to Modumate.app
+    mv "${UE_OUTPUT_APP}" "${folder}/Modumate.app"
+    UE_OUTPUT_APP="${folder}/Modumate.app"
+
+    #move the binary in the same way
+    mv "${UE_OUTPUT_APP}/Contents/MacOS/Modumate-Mac-Shipping" "${UE_OUTPUT_APP}/Contents/MacOS/Modumate"
+
+    #Update the reference to the plist so it points to the
+    # renamed app bundle
+    INFO_PLIST="${UE_OUTPUT_APP}/Contents/Info.plist"
+    plutil -replace CFBundleName -string Modumate $INFO_PLIST
+    plutil -replace CFBundleExecutable -string Modumate $INFO_PLIST
+
+    # For every dylib in the app bundle (excluding CEFs), change the rpaths
+    # dirname gives us a path up to current pwd, so make sure we've cd'd into the package before adding rpaths    
+    cd "${UE_OUTPUT_APP}/Contents/"
+    find "./UE4/Engine/Binaries/ThirdParty/PhysX3" -name "*.dylib" -type f -exec sh -c 'install_name_tool -delete_rpath "/Volumes/Work/Perforce/UE4/Engine/Binaries/ThirdParty/PhysX3/Mac" "$1"' sh {} \;
+    find "./UE4" -name "*.dylib" -type f -exec sh -c 'install_name_tool -add_rpath $(dirname "$1") "$1"' sh {} \;
+    cd $oldDir
+
+fi
+
+#-------------------------------------
+# Add URL Scheme to plist
+plutil -replace CFBundleURLTypes -json '[]' $INFO_PLIST
+plutil -replace CFBundleURLTypes.0 -json '{}' $INFO_PLIST
+
+plutil -insert CFBundleURLTypes.0.CFBundleTypeRole -string "Viewer" $INFO_PLIST
+plutil -insert CFBundleURLTypes.0.CFBundleURLName -string "com.Modumate.Modumate" $INFO_PLIST
+plutil -insert CFBundleURLTypes.0.CFBundleURLSchemes -json '["mdmt"]' $INFO_PLIST
+plutil -replace NSMicrophoneUsageDescription -string "Uses the microphone for chat" $INFO_PLIST
+
+#-------------------------------------
+# Modify PLIST with correct version and add provisionprofile
 if [[ ! -z "$BUILD_VERSION" ]]; then
     if [[ ! -z "$ENGINE_VERSION" ]]; then
         if [[ -f "${INFO_PLIST}" ]]; then
-            echo "${highlight}Updating Info.plist${normal}"
+            echo "${highlight}Updating Info.plist version${normal}"
             cd ${MODUMATE_GIT_PATH}
             sedcom="sed -I b 's/"${ENGINE_VERSION}"/"${BUILD_VERSION}"/' ${INFO_PLIST}"
             echo $sedcom
